pyproject.toml
[project]
name = "youtube-to-obsidian"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "notebooklm-py[browser]>=0.3.4",
    "pyyaml>=6.0.3",
    "yt-dlp>=2026.3.3",
]
